refactor: move `Eval` to its own component
